CODE/homepass_per_odp/requirements.txt
pandas
openpyxl
paramiko
scp
python-dotenv